Dynamic Modules and RibbonButton now use a Description.qml for menu/name/etc (#4072)
Partially implements https://github.com/jasp-stats/INTERNAL-jasp/issues/686

Generating markdown help from QML and R (#4076)

* Generating markdown help from QML and R

Info now present on all jaspObjects and all jaspControls.
Implements https://github.com/jasp-stats/INTERNAL-jasp/issues/740
But the formatting can still use some work.

All filled info will be collected when the help is opened and will be updated whenever anything changes (aka translation).
For each qml element that has info but not a manually set "toolTip" the info will also be shown as a tooltip directly to give people easy access to the docs.
The info set in R is transmitted through the "meta" field and is coupled with the title of each object and displayed beneath the qml help.
This info could also be shown in between the results but I don't think that needs to be done right now.

Also set some info on Summary Statistics Bayesian Binomial Test to get a feel for how it will look, so to see an example of this: turn on developers mode and open that analysis.

* Forgot qsTr()

* Add qsTr to BayesFactorType.qml

* Moved the controlType to friendly name to the virtual friendlyName function on JASPControlWrapper

Wrong quote used in QML info property (#4079)

Windows decided that those urls are probably network resources.
I guess because they start with "E:" which could be interpreted as a protocol...
Also some problems with loading a dynamic module on top of a non-dynamic one fixed

Ok Q_PROPERTY macro does not like it when I use typedeffed stuff in it.

Small Interface change for TabView (#4108)

Adds 'debug' to entries in Description.qml
- They will only be shown in developer mode
- Implements https://github.com/jasp-stats/INTERNAL-jasp/issues/895
- Also makes sure that any parsing errors and stuff like that in Description.qml are shown in a msgbox to the user...

Fix spelling error